services/embedding_service/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
torch>=2.1.0
torchvision>=0.16.0
clip-by-openai>=1.0

# Image processing
Pillow>=10.0.0
numpy==1.24.3

# HTTP client for testing
httpx==0.25.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
